Moved commands into separate file
