+++ blended-updater/src/main/scala/blended/updater/internal/BlendedUpdaterActivator.scala
@@ -1,40 +1,22 @@
 package blended.updater.internal
 
 import java.io.File
-import java.io.FileOutputStream
-import java.io.PrintStream
-import java.util.UUID
-import java.util.concurrent.TimeUnit.SECONDS
-import domino.DominoActivator
-import scala.concurrent.Await
-import scala.concurrent.duration.HOURS
-import scala.concurrent.duration.MINUTES
+
 import scala.reflect.runtime.universe
 import scala.util.Failure
 import scala.util.Success
-import scala.util.Try
+
 import org.osgi.framework.ServiceRegistration
+
 import com.typesafe.config.ConfigFactory
-import com.typesafe.config.ConfigRenderOptions
-import akka.actor.ActorRef
-import akka.actor.ActorSystem
-import akka.pattern.ask
-import akka.util.Timeout
-import blended.akka.{ ActorSystemWatching, OSGIActorConfig }
-import blended.launcher.config.LauncherConfig
+
+import blended.akka.ActorSystemWatching
 import blended.updater.Updater
-import blended.updater.Updater.RuntimeConfigActivated
-import blended.updater.Updater.RuntimeConfigActivationFailed
-import blended.updater.Updater.RuntimeConfigAdded
-import blended.updater.Updater.RuntimeConfigAdditionFailed
-import blended.updater.Updater.RuntimeConfigStaged
-import blended.updater.Updater.RuntimeConfigStagingFailed
-import blended.updater.config.ConfigConverter
+import blended.updater.UpdaterConfig
 import blended.updater.config.ConfigWriter
 import blended.updater.config.ProfileLookup
 import blended.updater.config.RuntimeConfig
-import blended.updater.UpdaterConfig
-import blended.updater.config.LocalRuntimeConfig
+import domino.DominoActivator
 
 case class UpdateEnv(
   launchedProfileName: String,
@@ -50,9 +32,6 @@ class BlendedUpdaterActivator extends DominoActivator with ActorSystemWatching {
   whenBundleActive {
     whenActorSystemAvailable { cfg =>
       log.info(s"About to setup ${getClass()}")
-
-      //      val configDir = cfg.idSvc.getContainerContext().getContainerConfigDirectory()
-      //      val installDir = new File(cfg.idSvc.getContainerContext().getContainerDirectory(), "profiles").getAbsoluteFile()
 
       val restartFrameworkAction = { () =>
         val frameworkBundle = bundleContext.getBundle(0)
@@ -97,11 +76,21 @@ class BlendedUpdaterActivator extends DominoActivator with ActorSystemWatching {
             )
           )
 
+          def registerCommands(srv: AnyRef, cmds: Seq[(String, String)]): ServiceRegistration[Object] = {
+            val (commands, descriptions) = cmds.unzip
+            srv.providesService[Object](
+              "osgi.command.scope" -> "blended.updater",
+              "osgi.command.function" -> commands.toArray,
+              "blended.osgi.command.description" -> descriptions.toArray
+            )
+
+          }
+
+//          val osgiCommands = new OsgiCommands(bundleContext)
+//          registerCommands(osgiCommands, osgiCommands.commands)
+
           val commands = new Commands(actor, Some(updateEnv))(cfg.system)
-          commands.providesService[Object](
-            "osgi.command.scope" -> "blended.updater",
-            "osgi.command.function" -> commands.commands
-          )
+          registerCommands(commands, commands.commandsWithDescription)
 
       }
     }
@@ -117,8 +106,8 @@ class BlendedUpdaterActivator extends DominoActivator with ActorSystemWatching {
     val pProfileDir = Option(props.getProperty(RuntimeConfig.Properties.PROFILE_DIR))
     Some(
       UpdateEnv(
-        launchedProfileName = pName.get, 
-        launchedProfileVersion = pVersion.get, 
+        launchedProfileName = pName.get,
+        launchedProfileVersion = pVersion.get,
         launchProfileLookupFile = pProfileLookupFile.map(f => new File(f)),
         profilesBaseDir = new File(pProfilesBaseDir.get),
         launchedProfileDir = pProfileDir.map(f => new File(f))
@@ -135,87 +124,3 @@ class BlendedUpdaterActivator extends DominoActivator with ActorSystemWatching {
 
 }
 
-class Commands(updater: ActorRef, env: Option[UpdateEnv])(implicit val actorSystem: ActorSystem) {
-
-  val commands = Array("show", "add", "stage", "activate", "convertLauchConfigToRuntimeConfig")
-
-  def show(): AnyRef = {
-    implicit val timeout = Timeout(5, SECONDS)
-    val configs = Await.result(
-      ask(updater, Updater.GetRuntimeConfigs(UUID.randomUUID().toString())).mapTo[Updater.RuntimeConfigs],
-      timeout.duration)
-
-    def format(config: LocalRuntimeConfig): String = {
-      val activeSuffix = env match {
-        case Some(e) if e.launchedProfileName == config.runtimeConfig.name && e.launchedProfileVersion == config.runtimeConfig.version => " [active]"
-        case _ => ""
-      }
-      s"${config.runtimeConfig.name}-${config.runtimeConfig.version}${activeSuffix}"
-    }
-
-    "staged: " + configs.staged.map(format).mkString(", ") + "\n" +
-      "pending: " + configs.pending.map(format).mkString(", ") + "\n" +
-      "invalid: " + configs.invalid.map(format).mkString(", ")
-  }
-
-  def add(file: File): AnyRef = {
-    val config = ConfigFactory.parseFile(file).resolve()
-    val runtimeConfig = RuntimeConfig.read(config).get
-    println("About to add: " + runtimeConfig)
-
-    implicit val timeout = Timeout(5, SECONDS)
-    val reqId = UUID.randomUUID().toString()
-    Await.result(
-      ask(updater, Updater.AddRuntimeConfig(reqId, runtimeConfig)), timeout.duration) match {
-        case RuntimeConfigAdded(`reqId`) =>
-          "Added: " + runtimeConfig
-        case RuntimeConfigAdditionFailed(`reqId`, error) =>
-          "Failed: " + error
-        case x =>
-          "Error: " + x
-      }
-  }
-
-  def stage(name: String, version: String): AnyRef = {
-    implicit val timeout = Timeout(1, HOURS)
-    val reqId = UUID.randomUUID().toString()
-    Await.result(
-      ask(updater, Updater.StageRuntimeConfig(reqId, name, version)), timeout.duration) match {
-        case RuntimeConfigStaged(`reqId`) =>
-          "Staged: " + name + " " + version
-        case RuntimeConfigStagingFailed(`reqId`, reason) =>
-          "Failed: " + reason
-        case x =>
-          "Error: " + x
-      }
-  }
-
-  def activate(name: String, version: String): AnyRef = {
-    env match {
-      case Some(UpdateEnv(_, _, Some(lookupFile), _, _)) =>
-        implicit val timeout = Timeout(5, MINUTES)
-        val reqId = UUID.randomUUID().toString()
-        Await.result(
-          ask(updater, Updater.ActivateRuntimeConfig(reqId, name, version)), timeout.duration) match {
-            case RuntimeConfigActivated(`reqId`) =>
-              "Activated: " + name + " " + version
-            case RuntimeConfigActivationFailed(`reqId`, reason) =>
-              "Failed: " + reason
-            case x =>
-              "Error: " + x
-          }
-      case _ =>
-        sys.error("No updateable environment detected. No profile lookup file defined.")
-    }
-  }
-
-  def convertLauchConfigToRuntimeConfig(launcherConfig: File, runtimeConfig: File): Unit = {
-    val lc = LauncherConfig.read(launcherConfig)
-    val rc = ConfigConverter.launcherConfigToRuntimeConfig(lc, "???")
-    val rendered = RuntimeConfig.toConfig(rc).root().render(ConfigRenderOptions.concise().setFormatted(true).setJson(false))
-    val os = new PrintStream(new FileOutputStream(runtimeConfig))
-    os.print(rendered)
-    os.close()
-  }
-
-}
